+++ be/src/util/murmur_hash3.cpp
@@ -10,7 +10,7 @@
 #include "murmur_hash3.h"
 
 #include "vec/common/unaligned.h"
-
+#include "common/compile_check_begin.h"
 //-----------------------------------------------------------------------------
 // Platform-specific functions and macros
 
@@ -89,7 +89,7 @@ FORCE_INLINE uint64_t fmix64(uint64_t k) {
 
 void murmur_hash3_x86_32(const void* key, int64_t len, uint32_t seed, void* out) {
     const uint8_t* data = (const uint8_t*)key;
-    const int nblocks = len / 4;
+    const int nblocks = (int)len / 4;
 
     uint32_t h1 = seed;
 
@@ -439,7 +439,7 @@ void murmur_hash3_x64_128(const void* key, const int len, const uint32_t seed, v
 
 void murmur_hash3_x64_64(const void* key, const int64_t len, const uint64_t seed, void* out) {
     const uint8_t* data = (const uint8_t*)key;
-    const int nblocks = len / 8;
+    const int nblocks = (int)len / 8;
     uint64_t h1 = seed;
 
     const uint64_t c1 = BIG_CONSTANT(0x87c37b91114253d5);
@@ -503,5 +503,5 @@ void murmur_hash3_x64_64(const void* key, const int64_t len, const uint64_t seed
 
     ((uint64_t*)out)[0] = h1;
 }
-
+#include "common/compile_check_end.h"
 //-----------------------------------------------------------------------------
